refactor(html): replace html for test html source
